FEAT: Cleanup on a few warnings on casting

+++ src/Seagate/CScsi_Farm_Log.cpp
@@ -2703,22 +2703,22 @@ eReturnValues CSCSI_Farm_Log::print_Enviroment_Statistics_Page_07(JSONNODE *mast
     }
     json_set_name(pageInfo, temp.str().c_str());
 
-    double TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.average12v)) * static_cast<double>(0.001F));
+    double TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.average12v)) * 0.001L);
     set_json_float_With_Status(pageInfo, "Current 12 volts", TempValue, vFarmFrame.at(page).envStatPage07.average12v, m_showStatusBits);
     if (m_MinorRev > 9 || m_showStatic == true)
     {
-        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.min12v)) * static_cast<double>(0.001F));
+        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.min12v)) * 0.001L);
         set_json_float_With_Status(pageInfo, "Minimum 12 volts", TempValue, vFarmFrame.at(page).envStatPage07.min12v, m_showStatusBits);
-        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.max12v)) * static_cast<double>(0.001F));
+        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.max12v)) * 0.001L);
         set_json_float_With_Status(pageInfo, "Maximum 12 volts", TempValue, vFarmFrame.at(page).envStatPage07.max12v, m_showStatusBits);
     }
-    TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.average5v)) * static_cast<double>(0.001F));
+    TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.average5v)) * 0.001L);
     set_json_float_With_Status(pageInfo, "Current 5 volts", TempValue, vFarmFrame.at(page).envStatPage07.average5v, m_showStatusBits);
     if (m_MinorRev > 9 || m_showStatic == true)
     {
-        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.min5v)) * static_cast<double>(0.001F));
+        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.min5v)) * 0.001L);
         set_json_float_With_Status(pageInfo, "Minimum 5 volts", TempValue, vFarmFrame.at(page).envStatPage07.min5v, m_showStatusBits);
-        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.max5v)) * static_cast<double>(0.001F));
+        TempValue = static_cast<double>(M_Word0(check_Status_Strip_Status(vFarmFrame.at(page).envStatPage07.max5v)) * 0.001L);
         set_json_float_With_Status(pageInfo, "Maximum 5 volts", TempValue, vFarmFrame.at(page).envStatPage07.max5v, m_showStatusBits);
     }
 
@@ -3075,7 +3075,7 @@ eReturnValues CSCSI_Farm_Log::print_Head_Information(eSASLogPageTypes type, JSON
             {
                 for (loopCount = 0; loopCount < m_heads; ++loopCount)
                 {
-                    printf("\tCurrent H2SAT trimmed mean bits in error Zone 0 for Head %2" PRIu32":   %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STTrimmedbyHeadZone0.headValue[loopCount])), 16)) * 0.10F);
+                    printf("\tCurrent H2SAT trimmed mean bits in error Zone 0 for Head %2" PRIu32":   %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STTrimmedbyHeadZone0.headValue[loopCount])), 16)) * 0.10);
                 }
             }
             sas_Head_Float_Data(headPage, "Current H2SAT trimmed mean bits in error Zone 0", WORD0, 0.10, vFarmFrame.at(page).currentH2STTrimmedbyHeadZone0.headValue, m_heads, m_showStatusBits, m_showStatic);
@@ -3085,7 +3085,7 @@ eReturnValues CSCSI_Farm_Log::print_Head_Information(eSASLogPageTypes type, JSON
             {
                 for (loopCount = 0; loopCount < m_heads; ++loopCount)
                 {
-                    printf("\tCurrent H2SAT trimmed mean bits in error Zone 1 for Head %2" PRIu32":   %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STTrimmedbyHeadZone1.headValue[loopCount])), 16)) * 0.10F);
+                    printf("\tCurrent H2SAT trimmed mean bits in error Zone 1 for Head %2" PRIu32":   %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STTrimmedbyHeadZone1.headValue[loopCount])), 16)) * 0.10);
                 }
             }
             sas_Head_Float_Data(headPage, "Current H2SAT trimmed mean bits in error Zone 1", WORD0, 0.10, vFarmFrame.at(page).currentH2STTrimmedbyHeadZone1.headValue, m_heads, m_showStatusBits, m_showStatic);
@@ -3095,7 +3095,7 @@ eReturnValues CSCSI_Farm_Log::print_Head_Information(eSASLogPageTypes type, JSON
             {
                 for (loopCount = 0; loopCount < m_heads; ++loopCount)
                 {
-                    printf("\tCurrent H2SAT trimmed mean bits in error Zone 2 for Head %2" PRIu32":   %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STTrimmedbyHeadZone2.headValue[loopCount])), 16)) * 0.10F);
+                    printf("\tCurrent H2SAT trimmed mean bits in error Zone 2 for Head %2" PRIu32":   %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STTrimmedbyHeadZone2.headValue[loopCount])), 16)) * 0.10);
                 }
             }
             sas_Head_Float_Data(headPage, "Current H2SAT trimmed mean bits in error Zone 2", WORD0, 0.10, vFarmFrame.at(page).currentH2STTrimmedbyHeadZone2.headValue, m_heads, m_showStatusBits, m_showStatic);
@@ -3105,7 +3105,7 @@ eReturnValues CSCSI_Farm_Log::print_Head_Information(eSASLogPageTypes type, JSON
             {
                 for (loopCount = 0; loopCount < m_heads; ++loopCount)
                 {
-                    printf("\tCurrent H2SAT iterations to converge Test Zone 0 for Head %2" PRIu32":  %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STIterationsByHeadZone0.headValue[loopCount])), 16) )* 0.10F);
+                    printf("\tCurrent H2SAT iterations to converge Test Zone 0 for Head %2" PRIu32":  %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STIterationsByHeadZone0.headValue[loopCount])), 16) )* 0.10);
                 }
             }
             sas_Head_Float_Data(headPage, "Current H2SAT iterations to converge Test Zone 0", WORD0, 0.10, vFarmFrame.at(page).currentH2STIterationsByHeadZone0.headValue, m_heads, m_showStatusBits, m_showStatic);
@@ -3115,7 +3115,7 @@ eReturnValues CSCSI_Farm_Log::print_Head_Information(eSASLogPageTypes type, JSON
             {
                 for (loopCount = 0; loopCount < m_heads; ++loopCount)
                 {
-                    printf("\tCurrent H2SAT iterations to converge Test Zone 1 for Head %2" PRIu32":  %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STIterationsByHeadZone1.headValue[loopCount])), 16)) * 0.10F);
+                    printf("\tCurrent H2SAT iterations to converge Test Zone 1 for Head %2" PRIu32":  %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STIterationsByHeadZone1.headValue[loopCount])), 16)) * 0.10);
                 }
             }
             sas_Head_Float_Data(headPage, "Current H2SAT iterations to converge Test Zone 1", WORD0, 0.10, vFarmFrame.at(page).currentH2STIterationsByHeadZone1.headValue, m_heads, m_showStatusBits, m_showStatic);
@@ -3125,7 +3125,7 @@ eReturnValues CSCSI_Farm_Log::print_Head_Information(eSASLogPageTypes type, JSON
             {
                 for (loopCount = 0; loopCount < m_heads; ++loopCount)
                 {
-                    printf("\tCurrent H2SAT iterations to converge Test Zone 2 for Head %2" PRIu32":  %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STIterationsByHeadZone2.headValue[loopCount])), 16))* 0.10F);
+                    printf("\tCurrent H2SAT iterations to converge Test Zone 2 for Head %2" PRIu32":  %0.3f \n", loopCount, static_cast<double>(check_for_signed_int(M_WordInt0(check_Status_Strip_Status(vFarmFrame.at(page).currentH2STIterationsByHeadZone2.headValue[loopCount])), 16))* 0.10);
                 }
             }
             sas_Head_Float_Data(headPage, "Current H2SAT iterations to converge Test Zone 2", WORD0, 0.10, vFarmFrame.at(page).currentH2STIterationsByHeadZone2.headValue, m_heads, m_showStatusBits, m_showStatic);
